requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
idna==3.6
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytz==2023.3.post1
pywatchman==1.4.1
requests==2.31.0
sqlparse==0.4.4
urllib3==2.2.1
